Switch from openssl to libressl
Make openssl a fake package that depends on libressl
